+++ projects/ng-smooth-carousel/src/lib/carousel-config.interface.ts
@@ -1,10 +1,10 @@
-export type ScrollSize = 'xs' | 'sm' | 'md' | 'lg' | 'xl' | '2xl' | '3xl' | '4xl' | '5xl' | '6xl' | '7xl' | '8xl' | '9xl' | '10xl' | 'full';
+export type ScrollSize = 'xs' | 'sm' | 'md' | 'lg' | 'xl' | '2xl' | '3xl' | '4xl' | '5xl' | '6xl' | '7xl' | '8xl' | '9xl' | '10xl';
 
 export type NavButtonShape = 'circle' | 'square' | 'rounded';
 
 export interface NavigationStyle {
-  nextButton?: Record<string, string>;
-  prevButton?: Record<string, string>;
+  nextButton?: Record<string, string> | ButtonStyle;
+  prevButton?: Record<string, string> | ButtonStyle;
   buttonShape?: NavButtonShape;
   icons?: {
     next?: string;
@@ -40,10 +40,13 @@ export interface CarouselConfig {
   autoplayDelay?: string;
   loop?: boolean;
   scrollSize?: string;
-  enableSearch?: boolean;
-  searchPlaceholder?: string;
-  searchModalTitle?: string;
-  searchStyle?: SearchStyle;
+  /**
+   * @description Enable search functionality (COMING SOON)
+   */
+  // enableSearch?: boolean;
+  // searchPlaceholder?: string;
+  // searchModalTitle?: string;
+  // searchStyle?: SearchStyle;
   animate?: boolean;
   responsive?: boolean;
   enableOneItemScroll?: boolean;
@@ -57,4 +60,11 @@ export interface ButtonStyle {
   fontSize?: string;
   border?: string;
   boxShadow?: string;
+  top?: string;
+  bottom?: string;
+  left?: string;
+  right?: string;
+  zIndex?: string;
+  transform?: string;
+  position?: string;
 }