+++ projects/ng-smooth-carousel/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [14.1.0] - 2025-04-20
+
+### Added
+- Added individual position styling for prev and next buttons
+- Each button can now have its own position properties (top, bottom, left, right, position, transform, zIndex)
+- Position properties can be specified directly in the nextButton and prevButton objects
+
 ## [14.0.9] - 2025-04-18
 
 ### Added